requirements.txt
click==7.1.2
dnspython==1.16.0
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
pymongo==3.11.3
python-dotenv==0.15.0
python-snappy==0.6.0
Werkzeug==1.0.1
winkerberos==0.7.0
gunicorn